chore(storybook): cleaner preview file
